[UPD] refactor
