+++ pymap/monitor/monitor.py
@@ -3,8 +3,8 @@ from time import sleep
 import datetime
 
 from pymap.includes.config import alert_envs
-from pymap.methods.marker_methods import MarkerMethods
-from pymap.methods.makalu_api_methods import MakaluApiMethods
+from pymap.process.methods.marker.marker_methods import MarkerMethods
+from pymap.rpc.methods.makalu_api_methods import MakaluApiMethods
 from pymap.tools.general import General
 from pymap.monitor.alerts.send_alerts import Alerts
 from pymap.monitor.includes.monitor_setup import times_sent
