+++ run_monitor.py
@@ -1,5 +1,3 @@
-from code import interact
-from enum import auto
 from pymap.includes.config import envs
 from pymap.monitor.monitor import Monitor
